microservices/requirements.txt
amqp==2.5.2
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
eventlet==0.25.2
flasgger==0.9.4
Flask==1.1.2
greenlet==0.4.15
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.8
MarkupSafe==1.1.1
marshmallow==3.5.1
mistune==0.8.4
mock==4.0.2
monotonic==1.5
nameko==2.12.0
path==13.1.0
path.py==12.4.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dotenv==0.13.0
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
